+++ CHANGELOG.md
@@ -17,6 +17,11 @@
   - Buffed Universal VIP Pass Chance from 45% to 55%
   - Nerfed Red Spring Water to have diminishing returns scaling with buff count
   - Improved Amber Knife VFX and SFX
+- Interactables
+  - Added Delta Construct and Gamma Construct to Shrine of the Future
+  - Added a few more buffs to Corrupted Catalyst
+  - Buffed Shrine of The Future enemy stats from 50% to 60% 
+  - Fixed Shrine of Ruin spawning a standard teleporter on stage 5
 
 ## 1.0.7 Patch Notes
 - VOL-T